Move Xamarin.Forms-related code into LoginApp.Forms
