+++ samples/xamarin-forms/LoginApp/LoginApp.Android/LoginApp.Android.csproj
@@ -92,9 +92,9 @@
     <Folder Include="Resources\drawable-xxxhdpi\" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\LoginApp\LoginApp.csproj">
+    <ProjectReference Include="..\LoginApp.Forms\LoginApp.Forms.csproj">
       <Project>{A4B47D0D-AA6A-4698-A8FE-612ADE83BBFC}</Project>
-      <Name>LoginApp</Name>
+      <Name>LoginApp.Forms</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
