+++ samples/xamarin-forms/LoginApp/LoginApp.Forms/App.xaml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <Application xmlns="http://xamarin.com/schemas/2014/forms" xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="LoginApp.App">
+             x:Class="LoginApp.Forms.App">
     <Application.Resources>
         <ResourceDictionary>
             <Style x:Key="ErrorLabelStyle" TargetType="Label">
