+++ samples/xamarin-forms/LoginApp/LoginApp.Forms/LICENSE
