+++ samples/xamarin-forms/LoginApp/LoginApp.Forms/LoginApp.Forms.csproj
@@ -1,0 +1,35 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+    <DebugType>pdbonly</DebugType>
+    <DebugSymbols>true</DebugSymbols>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Acr.UserDialogs" Version="7.0.4" />
+    <PackageReference Include="ReactiveUI" Version="9.11.3" />
+    <PackageReference Include="ReactiveUI.Fody" Version="9.11.3" />
+    <PackageReference Include="ReactiveUI.Validation" Version="1.0.3" />
+    <PackageReference Include="ReactiveUI.XamForms" Version="9.11.3" />
+    <PackageReference Include="Xamarin.Forms" Version="3.6.0.220655" />
+  </ItemGroup>
+  
+  <ItemGroup>
+    <ProjectReference Include="..\LoginApp\LoginApp.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Update="Views\SignUpView.xaml.cs">
+      <DependentUpon>SignUpPage.xaml</DependentUpon>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Update="Views\SignUpView.xaml">
+      <Generator />
+    </EmbeddedResource>
+  </ItemGroup>
+</Project>
