+++ samples/xamarin-forms/LoginApp/LoginApp.iOS/AppDelegate.cs
@@ -5,6 +5,7 @@
 // </copyright>
 
 using Foundation;
+using LoginApp.Forms;
 using UIKit;
 
 namespace LoginApp.IOS
