+++ samples/xamarin-forms/LoginApp/LoginApp/ViewModels/SignUpViewModel.cs
@@ -7,7 +7,6 @@
 using System;
 using System.Reactive;
 using Acr.UserDialogs;
-using LoginApp.ViewModels.Abstractions;
 using ReactiveUI;
 using ReactiveUI.Fody.Helpers;
 using ReactiveUI.Validation.Abstractions;
